Httpd module:
Fixed a goof in redirect handlers.

Fixed which headers are accessed for logging.

Added a shim for web servers to provide a standard style mixin.

Added a shim for servers to participate with their own private methods 
prior and following mixin dispatch.

Added a hook for server to provide a local method for Headers