Adapted the Autodoc system to not embed a literal phrase that sak was using as a detector to see if a file was a manual page or not.
The autodoc manpage generator now accepts a substitution map as an argument
and will apply that substitution to boiler plate code added at the top of 
the file to declare the .man file a manual page. The same substitutions will
be performed on header and footer files as well.

Updated clay, httpd, and practcl build scripts to use the doctools resident
with Practcl