picoirc: remove custom message handling based on nicks
A general purpose library is not the place to perform #tcl specific
custom message handling. This should be done at the application level.

Removed the special handling for PRIVMSG to the current user's nick as
again it is up to the application to decide how to handle messages to the
current user and the application should received all valid fields.