ldap [B] Ticket [af927e957f]
Replaced <code>asnOctetString</code> in <code>starttls</code> and
<code>Whoami</code> with <code>asnEncodeString 80</code> to properly
signal "Extended Operation".

While detected with a MS ActiveDirectory server this is not an
MS-specific issue. Thus the unconditional change. It actually is not
even possible in the protocol to query the kind of server.

Bumped to 1.10.1

Thanks to Christian Nassau <nassau@nullhomotopie.de> for the report,
patience, and helping me with my questions.